물약, 총알 버그 수정

+++ Source/blockers/Private/BKChunkBase.h
@@ -31,7 +31,7 @@ struct FBlockk {
 	bool isCollapsible;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	uint8 CollapsibleType;
+	int32 CollapsibleType;
 };
 
 UCLASS(ABSTRACT)
